wip: charts with open-ended answers
