+++ apps/www/components/SurveyResults/charts/AcceleratorParticipationChart.tsx
@@ -1,4 +1,4 @@
-import { SurveyChart, buildWhereClause } from '../SurveyChart'
+import { SurveyChart, buildWhereClause, createCategoryAggregator } from '../SurveyChart'
 
 function generateAcceleratorParticipationSQL(activeFilters: Record<string, string>) {
   const whereClause = buildWhereClause(activeFilters, ['accelerator_participation IS NOT NULL'])
@@ -23,31 +23,15 @@ GROUP BY accelerator_clean
 ORDER BY total DESC;`
 }
 
-function transformAcceleratorParticipationData(data: any[]) {
-  // Raw data from Supabase: [{ accelerator_participation: 'YC' }, { accelerator_participation: 'Techstars' }, ...]
-  // Need to apply CASE logic and aggregate by counting occurrences
-  const acceleratorCounts: Record<string, number> = {}
-
-  data.forEach((row) => {
-    const accelerator = row.accelerator_participation
-
-    if (accelerator === 'Did not participate in an accelerator') {
-      return // Skip this case
-    }
-
-    let cleanAccelerator = accelerator
-    if (!['YC', 'Techstars', 'EF', '500 Global', 'Plug and Play', 'Antler'].includes(accelerator)) {
-      cleanAccelerator = 'Other'
-    }
-
-    acceleratorCounts[cleanAccelerator] = (acceleratorCounts[cleanAccelerator] || 0) + 1
-  })
-
-  // Convert to array format and sort by count descending
-  return Object.entries(acceleratorCounts)
-    .map(([accelerator_participation, total]) => ({ label: accelerator_participation, total }))
-    .sort((a, b) => b.total - a.total)
-}
+// Use the reusable helper for accelerator aggregation
+const getAcceleratorAggregates = createCategoryAggregator('accelerator_participation', [
+  'YC',
+  'Techstars',
+  'EF',
+  '500 Global',
+  'Plug and Play',
+  'Antler',
+])
 
 export function AcceleratorParticipationChart() {
   return (
@@ -56,7 +40,7 @@ export function AcceleratorParticipationChart() {
       targetColumn="accelerator_participation"
       filterColumns={['person_age', 'location', 'money_raised']}
       generateSQLQuery={generateAcceleratorParticipationSQL}
-      transformData={transformAcceleratorParticipationData}
+      customAggregateFunction={getAcceleratorAggregates}
     />
   )
 }
